+++ apps/web/src/app/wrong/create/components/steps/step-4.tsx
@@ -1,7 +1,4 @@
-"use client";
-
 import Image from "next/image";
-import { useCallback, useEffect, useMemo, useState } from "react";
 import TextField from "@/shared/components/text-field/text-field";
 import TextAreaField from "@/shared/components/text-area-field/text-area-field";
 import { Toggle } from "@/shared/components/toggle/toggle";
@@ -10,74 +7,13 @@ import { StepProps } from "@/app/wrong/create/page";
 import { TOGGLE_OPTIONS } from "@/app/wrong/create/constants/option-labels";
 import SampleImg from "@/shared/assets/images/wrong-sample.png";
 import * as s from "@/app/wrong/create/components/steps/step.css";
-
-type ToggleValue = "objective" | "subjective";
-
-type FormState = {
-  type: ToggleValue;
-  answerChoice: number | null;
-  answerText: string;
-  solutionText: string;
-};
+import {
+  useStep4Form,
+  type ToggleValue,
+} from "@/app/wrong/create/hooks/use-step4-form";
 
 const Step4 = ({ onNextEnabledChange }: StepProps) => {
-  const [form, setForm] = useState<FormState>({
-    type: "objective",
-    answerChoice: null,
-    answerText: "",
-    solutionText: "",
-  });
-
-  const updateForm = useCallback((patch: Partial<FormState>) => {
-    setForm((prev) => ({ ...prev, ...patch }));
-  }, []);
-
-  const isNextEnabled = useMemo(() => {
-    const hasAnswer =
-      form.type === "objective"
-        ? form.answerChoice !== null
-        : form.answerText.trim().length > 0;
-
-    const hasSolution = form.solutionText.trim().length > 0;
-
-    return hasAnswer && hasSolution;
-  }, [form]);
-
-  useEffect(() => {
-    onNextEnabledChange?.(isNextEnabled);
-  }, [isNextEnabled, onNextEnabledChange]);
-
-  const handleTypeChange = useCallback(
-    (nextType: ToggleValue) => {
-      updateForm({
-        type: nextType,
-        answerChoice: null,
-        answerText: "",
-      });
-    },
-    [updateForm]
-  );
-
-  const handleChoiceChange = useCallback(
-    (nextChoice: number) => {
-      updateForm({ answerChoice: nextChoice });
-    },
-    [updateForm]
-  );
-
-  const handleAnswerTextChange = useCallback(
-    (nextAnswerText: string) => {
-      updateForm({ answerText: nextAnswerText });
-    },
-    [updateForm]
-  );
-
-  const handleSolutionChange = useCallback(
-    (nextSolution: string) => {
-      updateForm({ solutionText: nextSolution });
-    },
-    [updateForm]
-  );
+  const { form, handlers } = useStep4Form(onNextEnabledChange);
 
   return (
     <div className={s.step4Container}>
@@ -97,7 +33,7 @@ const Step4 = ({ onNextEnabledChange }: StepProps) => {
 
             <Toggle<ToggleValue>
               value={form.type}
-              onValueChange={handleTypeChange}
+              onValueChange={handlers.handleTypeChange}
               options={TOGGLE_OPTIONS}
             />
           </div>
@@ -105,14 +41,14 @@ const Step4 = ({ onNextEnabledChange }: StepProps) => {
           {form.type === "objective" ? (
             <NumberChoice
               value={form.answerChoice}
-              onValueChange={handleChoiceChange}
+              onValueChange={handlers.handleChoiceChange}
             />
           ) : (
             <TextField
               fullWidth
               placeholder="정답을 입력해주세요."
               value={form.answerText}
-              onChange={(e) => handleAnswerTextChange(e.target.value)}
+              onChange={(e) => handlers.handleAnswerTextChange(e.target.value)}
             />
           )}
         </div>
@@ -123,7 +59,7 @@ const Step4 = ({ onNextEnabledChange }: StepProps) => {
             fullWidth
             placeholder="풀이를 입력해주세요."
             value={form.solutionText}
-            onChange={(e) => handleSolutionChange(e.target.value)}
+            onChange={(e) => handlers.handleSolutionChange(e.target.value)}
           />
         </div>
       </div>
